requirements.txt
discord.py-self==2.0.0
httpx==0.27.0
tenacity==8.4.1
python-dotenv==1.0.1
pydantic==1.10.15
pytest==8.2.2
ruff==0.5.5
flake8==7.1.1

